Do not recurse arbitrarily down the source directory for opam files
Instead, look only in the root directory of the project, and also
in the `duniverse/` directory.  The former strategy is what opam
does, and the latter is a convention.

This prevents failures with projects that have example or broken
opam files (for example, dune-configurator has tests that are
pinned and broken).