do not pin with a .dev version
The .dev version is unfortunately less than the `v0.13.0` method
used by the JS packages, and so we need to go higher.

The duniverse tool now stores the version of the dependency in
the directory name of the `duniverse/` vendor area, and so opam
picks up the original version. This also respects constraints
better than forcing everything to `.dev`.

Fixes the CI with ppx_sexp_conv and other JS libraries.