for duniverses, do not pin and run opam install
For duniverses, the state of the opam files in the local duniverse
may not be entirely consistent, and pinning them beyond the needs
of depext is unnecessary. The user will be running `dune build`
and so the test does just this, and runs the tests and doc build.

The `opam pin` is also reverted before the main build, since it
leads to hard to track down errors. The initial depext pin will
result in an _empty_ build directory since only the opam files are
copied over in the first stage.  If an opam update is not done
subsequently to trying to opam install the dependencies, then the
package installations fail. Opam seems to not always do an rsync
update upon install, so I ran into this several times.

The new strategy should leave normal opam projects unaffected,
beyond just unpinning and pinning one extra time for the root
package.