Pin packages as .dev again, unless they specify a version
If the package's directory name doesn't contain a dot then opam will
default to using the last known version, which is usually wrong. In
particular, if a multi-project repostory adds a new package with a
constraint `{ =version }` on an existing one, this will fail because
opam will pin the new package as "dev" but the old one with the version
of its last release.

This partially reverts 7380c74681. Duniverse projects should be
unaffected, because they always specify the version explicitly.