Run solver in the analysis phase
This is faster, and avoids the need to run separate Docker builds on
platforms for which there is no solution. More importantly, it allows us
to follow changes in opam-repository more closely, since updating that
no longer requires pulling a new set of base images and rebuilding
everything.

We reject old opam versions (1.x) because the solver doesn't understand
some fields (e.g. `available`) and it's better to encourage people to
upgrade, anyway.

We set the opam dev flag to true for `.dev` packages, which seems to
match what opam does.