Report status as pending during analysis
Previously, we waited until we had the list of builds before reporting
any build status (e.g. pending). At the moment that's not a problem,
because `platforms` is a constant, but in future when it comes from the
analysis step this would delay reporting the commit status until the
analysis phase was complete.

Spotted by kit-ty-kate.