Combine lint with new build code
This adds a new `Build.Spec.t` type to represent the build task to be
performed. This also allows us to pass just the part of the analysis
needed for each job. At the moment, opam jobs get everything and
duniverse jobs get nothing, but in future we'll probably want to pass
per-platform package selections here too.

This means that:

- The Docker pull for lint jobs is latched too.
- Lint jobs display the Dockerfile like the other jobs do.