Use git-unix to read opam-repository
Instead of checking out the whole of opam-repository on each analysis
run, just pass the hash to the solver. The solver maintains a bare
opam-repository clone and reads it directly.

We also parse all the opam files in opam-repository and keep them in
memory until the head commit changes, which should be faster when
doing lots of solves.