duniverse: do not pin the tool
The ocamlformat failure on the live CI is because the opam metadata
in the release branch is wrong. Since these tools are all guaranteed
to be in the opam repo, simply do not pin the tool while building
the binary.

This should fix the live CI which is failing with ocamlformat
trying to use base.v0.14.0 and thus not building.