Report statistics about the state of all master branches
Reporting states for all PRs and branches isn't very useful because they
may be stale, but the master branches should be a good indicator of
things that are supposed to be passing.

Also, we now track the set of active repositories. Before, we listed all
repositories for which there were entries in the database, even if they
are no longer being monitored. Now, we just list the ones that are still
active.