support 32-bit builds
This plumbs through an arch option in order to allow the selection
of 32bit builds.  Its only activated for i386 right now, but can
easily be extended to support other architectures in the cluster
pool as well.

It adds a Ocaml_ci.Variant.t to track the tuple of image and
architecture. This is necessary to track since the solver results
may be influenced by the architecture, and so we need to ensure
that it is represented in the id string passed out to the solver
service.