duniverse: use opam-tools plugin to get environment
The duniverse build previously hardcoded the set of tools to
build, which is obviously problematic when they get out of sync
(most notably for ocamlformat).  This now shifts the CI to using
the soon-to-be-released opam-tools plugin to get those tools
installed in a local switch.

I'll be refining this more, but this immediately unbreaks ocaml-ci
for realworldocaml/book, where the ocamlformat build has failed
for some time due to the old way of building it here.