Add support for building OCaml compiler repositories and variants
This adds support for parsing compiler variants (e.g. flambda) in
the Variant module, which in turns allows for the compilation of
OCaml compilers.

OCaml compiler repositories are detected (using the experimental
dune flag that signifies that the compiler dune rules are active).
We then use the base distros and variants defined for the trunk
compiler (in service/conf.ml) to generate Dockerfiles that run
the full test suite.

In order to prevent an explosion of tests for non-OCaml compilers
(i.e. normal repositories), we filter out all the non-standard
variants before solving for normal repositories.  We could bring
in opt-in support for (e.g. flambda or nnp) by looking for x-
fields in their opam files, but this is not implemented yet.