Add support for building OCaml variant versions
This adds support for parsing compiler variants (e.g. flambda) in
the Variant module.

The arch value is now explicitly specified, so the build modules can
choose whether to introduce a personality or not in their
Dockerfiles.

The configuration now uses Ocaml_version to setup the various
architecture choices.  This just adds flambda to the set of
tested defaults, using the latest stable compiler version.