lint-doc: copy project source into container before building docs
Fix https://github.com/ocurrent/ocaml-ci/issues/250. The implementation
of `Lint.doc_dockerfile` is now almost identical to that of
`Opam_build.dockerfile`, as it should be.

Also fixes an escaping issue in the use of "odoc>=1.5.0" which was
causing the creation of a file named "=1.5.0" inside the lint container.