monorepo: support exact compiler version
When the exact compiler version (down to the patch level) is available
in a base image, it is used. However, if it is not (for example the base
image has 4.10.2 but the project has 4.10.1 in its lockfile), it would
fail.

This adds a fallback mechanism for this situation by using the closest
base image (4.10 in the example) and creating a switch with the exact
version.

Closes #281