Fix solver worker failed: Unexpected end of input
In some cases, the C git client can create pack files that ocaml-git
cannot read. See mirage/ocaml-git#514, mirage/ocaml-git#515.

This is fixed in `carton.0.4.2` (an internal dependency of `ocaml-git`).
Since we do not directly use it in `ocaml-ci`, this upgrade is done
through a `conflicts` clause.

It can be removed once a version of `ocaml-git` is released with a bound
to `carton` greater than `0.4.2`.