test all packages with multicore OCaml if they are compatible
The ocaml-multicore/multicore-opam repo is now pulled in by default
for solving for all packages. This is safe to do since the overlay
repository guarantees that it only introduces packages that depend
on either `base-domains` or `base-effects`, and thus will not get
selected on vanilla compilers.  Any package that does not do this and
interferes with vanilla solutions is a bug, and will be fixed promptly
in multicore-opam.

Any package that is known-broken with multicore can indicate this
via a conflict with `base-depends`, and it will not be selected for
testing by ocaml-ci