Support opam-version variable
Closes #391

This adds `opam_version` to the set of worker vars. It is parsed from
the opam template `%{opam-version}`. Special support is taken when
creating the 0install context because the function `std_env` expands
this variable to the version of the opam library we are linking against.
This does seem to correspond to what we want here.

See https://github.com/ocaml-opam/opam-0install-solver/pull/36