For each variant, only build compatible packages
This checks each local .opam file for an explicit constraint on the
version of `ocaml`. If it doesn't match the platform, the package is
skipped.

This is useful when different packages in a repository work on different
versions of OCaml. For example, `prometheus` works on OCaml 4.03, but
`prometheus-app` requires 4.08. Previously, this prevented testing on
4.03 at all.