Only build multicore packages on multicore compilers
If a package depends on base-domains or base-effects then it requires a
multicore compiler. Don't try to test it on other platforms.

Many projects want to add a sub-package with support for effects on
OCaml 5.0, alongside their existing support for Lwt. This PR allows them
to do that without losing CI on all older compilers.