monorepo: do not crash if lockfiles share variant
Since #378, we create several builds when there are several lockfiles.
Unfortunately, if they use the same variant, this creates an ocurrent
graph that is invalid because several sibling nodes have the same label
(and this crashes the server).

To address this, this commit changes the labels when there are several
lockfiles and use the path to the lockfile.