Replace usage of socketpair for IPC
Everything works as expected on Linux, but the folling sections apply
on Windows.

Using a socket as standard input file descriptor is daring. Even if it
may be possible, Lwt will currently outright reject it. We can,
instead of passing a file descriptor to the spawned process, pass the
address of the UNIX domain socket as an argument so that the sub
process may open a socket itself and connect to that address.

Creating a filesystem entry with the same name as the socket seems
also to be a problem. We can't really test whether the socket exists
with `Sys.file_exists` (`_wstati64`), or `Cmdliner.Arg.file`, but we
can still unlink the socket.

Although Unix domain sockets might appear in the filesystem, it's all
a lie made by the kernel. Windows currently doesn't support abstract
socket, or unnamed sockets returned by `socketpair` (`socketpair`
isn't implemented).

UNIX domain sockets are supported on Windows 10 and OCaml 4.14.