Restore cancellation on opam file lookup
Now the opam file lookup is executed in a `Lwt_preemptive` thread and
should not block the main execution thread, but it is not cancellable
anymore.

Restore the possibility of cancelling the lookup by introducing a
signalling atomic. Regularly check the atomic: if cancellation was
requested, raise to exit early.