Exclude deb-12 x86_32 from builds with capnproto
Debian 12, "bookwork" is currently on a stable package repo, and
package versions are frozen. The frozen version of capnproto is
0.9.2. This version does not support large files on 32 bit arch,
and this can result in specious test and build failures.

These false positives have been giving us noisy red trunks for a long
while, but this fix should clean up our builds until we either drop
support for this target or the fix needed is backported to 0.9.2.